The meat of the program
up to this point, the program is able to save a folder with the
specified file name, and in the folder three files are saved: one for
each R(T), T(t), and SetP(t).  Next, I need to fix the graph so that it
doesn't plot in such an ugly way.